feat: Add event handler
