+++ src/bot.js
@@ -11,8 +11,6 @@ const client = new WelcomeBot({
     debug: process.env.NODE_ENV === "development",
 });
 
-const execute = require("./functions/execute");
-
 require("./db/connection");
 const getGuild = require("./db/functions/guild/getGuild");
 const dbAuditor = require("./db/functions/dbAuditor");
@@ -129,182 +127,6 @@ client.player
                 break;
         }
     });
-
-client.on("ready", () => {
-    // We logged in
-    if (client.debug)
-        client.logger.log(
-            `${client.user.tag}, ready to serve ${client.users.cache.size} users in ${client.guilds.cache.size} servers.`
-        );
-    client.onReady();
-});
-
-client.on("debug", (info) => {
-    if (
-        !info.match(/\b(?:heartbeat|token|connect)\b/gi) &&
-        client.debug &&
-        client.debugLevel > 0
-    )
-        client.logger.log(info, "debug", ["DISCORD"]);
-});
-
-client.on("rateLimit", (info) => {
-    client.logger.log("You are being rate limited!", "warn");
-    client.logger.log(JSON.stringify(info, null, 4), "warn");
-});
-
-client.on("guildMemberAdd", async (member) => {
-    // When a new member joins
-    require("./functions/greetUser")(member);
-    const guildDB = await getGuild(member.guild.id);
-    const t = client.i18next.getFixedT(guildDB.lang || "en-US");
-    if (guildDB.plugins.autorole.enabled) {
-        member.roles
-            .add(guildDB.plugins.autorole.role, "Autorole")
-            .catch(() => {});
-    }
-});
-
-client.on("guildMemberRemove", (member) => {
-    // When a member leaves or is kicked or is banned
-    require("./functions/sayGoodBye")(member);
-});
-
-client.on("guildCreate", (guild) => {
-    const lang = guild.preferredLocale || "en-US";
-    //Bot has been invited to a new guild
-    client.guildDbFuncs.addGuild(guild.id, lang);
-    if (guild.systemChannelID) {
-        guild.channels.cache
-            .get(guild.systemChannelID)
-            .send(
-                `Thank you for choosing this bot! To get started, type \`${
-                    client.config.defaultPrefix
-                }help\`\nJoin the support server: ${client.config.supportGuildInviteReal(
-                    client
-                )}`
-            )
-            .catch(() => {});
-    }
-    embed = new Embed({ color: "success", timestamp: true })
-        .setTitle(`:white_check_mark: Added to "${guild.name}"`)
-        .setDescription(`${guild.id}`)
-        .addField(
-            "Info",
-            `Shard: ${guild.shardId}\nOwner: <@${guild.ownerId}>\nMembers: ${guild.memberCount}`
-        );
-    client.channels.cache
-        .get(client.config.logsChannelId)
-        .send({ embeds: [embed] });
-});
-
-client.on("guildDelete", (guild) => {
-    //Bot has been kicked or banned in a guild
-    client.guildDbFuncs.removeGuild(guild.id);
-    embed = new Embed({ color: "error", timestamp: true })
-        .setTitle(`:x: Removed from "${guild.name}"`)
-        .setDescription(`${guild.id}`)
-        .addField(
-            "Info",
-            `Shard: ${guild.shardId}\nOwner: <@${guild.ownerId}>\nMembers: ${guild.memberCount}`
-        );
-    client.channels.cache
-        .get(client.config.logsChannelId)
-        .send({ embeds: [embed] });
-});
-
-client.on("interactionCreate", async (interaction) => {
-    if (!interaction.isCommand()) return;
-    const t = interaction.inGuild()
-        ? await getT(interaction.guild.id)
-        : client.i18next.getFixedT("en-US");
-    if (!client.application?.owner) await client.application?.fetch();
-    const { commandName: cmd } = interaction;
-    let guildDB;
-    if (interaction.inGuild() && interaction.channel.type !== "DM") {
-        guildDB = await getGuild(interaction.guild.id);
-    } else {
-        guildDB = { prefix: client.config.defaultPrefix, disabled: [] };
-    }
-    const userDB = await client.userDbFuncs.getUser(interaction.member.user.id);
-    const command = client.commands.enabled.get(cmd);
-    if (!command) return;
-    let preCheck = false;
-    preCheck = await command.preCheck(interaction, guildDB, t);
-    if (!preCheck) return;
-    command.run({ interaction, guildDB, userDB }, t).catch((err) => {
-        client.logger.log(`Error when executing ${command.name}`, "error", [
-            "CMDS",
-        ]);
-        console.log(err);
-        const embed = new Embed({ color: "error" })
-            .setTitle(t("errors:generic"))
-            .addField(
-                `Please report this to ${client.ownersTags.join(" OR ")}`,
-                "\u200b"
-            );
-        if (
-            client.config.ownerIds.includes(interaction.user.id) ||
-            client.config.staffIds.includes(interaction.user.id) ||
-            interaction.user.id === client.application?.owner.id
-        )
-            embed.addField("Error", `${err}`);
-        interaction.followUp({ embeds: [embed], ephemeral: true });
-    });
-    if (client.debug)
-        client.logger.log(`Executed ${command.name} command`, "debug");
-});
-
-client.on("messageCreate", async (message) => {
-    if (message.author.bot) return;
-    if (client.debug && client.debugLevel > 0)
-        client.logger.log("message event triggered", "debug");
-    //https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Operators#optional_chaining_operator
-    if (!client.application?.owner) await client.application?.fetch();
-    let guildDB;
-    if (message.guild && message.channel.type !== "DM") {
-        guildDB = await getGuild(message.guild.id);
-    } else {
-        guildDB = { prefix: client.config.defaultPrefix, disabled: [] };
-    }
-    if (message.channel?.partial) await message.channel.fetch();
-    if (message?.partial) await message.fetch();
-    if (client.debug && client.debugLevel > 0)
-        client.logger.log("running execute func", "debug");
-    try {
-        execute(message, guildDB);
-    } catch (e) {
-        client.logger.log(e, "error");
-    }
-    if (client.debug && client.debugLevel > 0)
-        client.logger.log("finished running execute func", "debug");
-    if (message.content.split(" ").length > 1) return;
-
-    const mentionRegex = new RegExp(`^(<@!?${message.client.user.id}>)\\s*`);
-    if (!mentionRegex.test(message.content)) return;
-    let reply = `Hi there, ${message.author}\nI am Welcome-Bot\nMy prefix is "${
-        guildDB.prefix
-    }"${message.guild ? " in this server." : ""}\nSend \`${
-        guildDB.prefix
-    }help\` to get help`;
-    if (message.guild) {
-        reply += `\nSend \`${guildDB.prefix}follow #channel\` where #channel is the channel you want to receive updates.`;
-    }
-    if (!message.reference) {
-        message.channel.sendTyping();
-        message.channel.send(reply);
-    } else {
-        message.channel.messages
-            .fetch(message.reference.messageID)
-            .then((msg) => {
-                if (msg.author.id != client.user.id) {
-                    message.channel.sendTyping();
-                    message.channel.send(reply);
-                }
-            })
-            .catch(console.error);
-    }
-});
 
 // Login
 client.login(process.env.DISCORD_TOKEN);
